Futz with Meson rpaths to get things working on all platforms.
Hoo boy. First, this patch adds a Meson `install_rpaths` setting to the
gdk-pixbuf shared libraries. Without this, the ppc64le build would fail in a
way that was super duper annoying to track down, and that I still don't quite
understand. The first error I got was "ELF load command alignment not
page-aligned", then when I fooled around I got segfaults in
gdk-pixbuf-query-loaders. But after much wailing and gnashing of teeth it
seems that this is all I need to get things working!

For `gdk-pixbuf-query-loaders` to actually work on ppc64le, I need to also set
the Meson `install_rpaths` setting for the loader dylibs -- otherwise the same
"load command alignment" error would occur on execution. But, that patch in
turn breaks the build on macOS due to `install_name_tool` being picky about
redundant rpaths, it seems. (Meson doesn't relay the error message from
`install_name_tool`, but if you run it manually on a macOS machine, you can
see the issue.) So we only conditionally apply that patch.